triple treats promo feature
git-svn-id: https://192.168.1.50:9880/easyshop_svn/trunk@4080 2929c1f8-612a-3548-b275-2a983e7c7391

+++ application/views/pages/product/product_upload_step2_view.php
@@ -853,8 +853,8 @@ $(document).ready(function(){
 
   $("#discountedP").on('change',function(){
     validateWhiteTextBox("#discountedP");
-    var disc_price = parseInt($(this).val());
-    var base_price = parseInt($("#prod_price").val().replace(/,/g,''));
+    var disc_price = parseFloat($(this).val());
+    var base_price = parseFloat($("#prod_price").val().replace(/,/g,''));
     var sum = ((base_price - disc_price) / base_price) * 100;
     sum = sum.toFixed(4);
     if(disc_price > base_price){
